+++ packages/blade/src/components/Radio/Radio.tsx
@@ -5,11 +5,11 @@ import type { OnChange } from './useRadio';
 import { useRadio } from './useRadio';
 import { RadioIcon } from './RadioIcon/RadioIcon';
 import { useRadioGroupContext } from './RadioGroup/RadioContext';
-import { FormHintText } from '~components/Form/FormHintText';
+import { SelectorLabel } from '~components/Form/Selector/SelectorLabel.web';
 import Box from '~components/Box';
-import { CheckboxLabel as RadioLabel } from '~components/Checkbox/CheckboxLabel';
+import { SelectorTitle } from '~components/Form/Selector/SelectorTitle';
+import { SelectorSupportText } from '~components/Form/Selector/SelectorSupportText';
 import { CheckboxInput as RadioInput } from '~components/Checkbox/CheckboxInput';
-import { CheckboxLabelText as RadioLabelText } from '~components/Checkbox/CheckboxLabelText';
 
 type RadioProps = {
   /**
@@ -70,7 +70,7 @@ const Radio = ({ value, children, helpText, isDisabled }: RadioProps): React.Rea
   });
 
   return (
-    <RadioLabel inputProps={state.isReactNative ? inputProps : {}}>
+    <SelectorLabel inputProps={state.isReactNative ? inputProps : {}}>
       <RadioInput
         isChecked={state.isChecked}
         isDisabled={_isDisabled}
@@ -79,14 +79,10 @@ const Radio = ({ value, children, helpText, isDisabled }: RadioProps): React.Rea
       />
       <RadioIcon isChecked={state.isChecked} isDisabled={_isDisabled} isNegative={hasError} />
       <Box>
-        <RadioLabelText>{children}</RadioLabelText>
-        {showHelpText && (
-          <FormHintText id={ids?.helpTextId} variant="help">
-            {helpText}
-          </FormHintText>
-        )}
+        <SelectorTitle>{children}</SelectorTitle>
+        {showHelpText && <SelectorSupportText id={ids?.helpTextId}>{helpText}</SelectorSupportText>}
       </Box>
-    </RadioLabel>
+    </SelectorLabel>
   );
 };