+++ packages/blade/src/components/Radio/RadioGroup/useRadioGroup.ts
@@ -59,7 +59,7 @@ const useRadioGroup = ({
           return;
         }
 
-        setValue(value);
+        setValue(() => value);
       },
       removeValue() {
         if (isDisabled) {