+++ packages/blade/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @razorpay/blade
 
+## 0.13.4
+
+### Patch Changes
+
+- 2778973: chore: add appropriate types for onClick of Button & BaseButton
+
+## 0.13.3
+
+### Patch Changes
+
+- 3ea6d6c: fix(blade): fixes checkbox helptext and errortext alignment for individual checkboxes
+
+## 0.13.2
+
+### Patch Changes
+
+- 17b2c71: fix: button styling for native
+
+## 0.13.1
+
+### Patch Changes
+
+- 985f82a: refactor: use Box component on BaseButton
+
 ## 0.13.0
 
 ### Minor Changes